% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClustMMDD.R
\docType{data}
\name{exModelKS}
\alias{exModelKS}

\title{An example of \code{\linkS4class{modelKS}}.}

\format{An instance of \code{\linkS4class{modelKS}}.}

\description{
  An example of a set of parameters given by an instance of \code{\linkS4class{modelKS}}.
}

\author{
Wilson Toussile
}

\seealso{
  \code{\linkS4class{modelKS}}
}

\examples{
data(exModelKS)
slotNames("modelKS")
head(exModelKS["membershipProba"])
exModelKS["mixingProportions"]
exModelKS
}

\references{
\itemize{
		\item  \href{http://projecteuclid.org/euclid.ejs/1379596773}{Dominique Bontemps and Wilson Toussile (2013)} : Clustering and 	variable selection for categorical multivariate data. Electronic Journal of Statistics, Volume 7, 2344-2371, ISSN.
		
		\item 	\href{http://link.springer.com/article/10.1007\%2Fs11634-009-0043-x}{Wilson Toussile and Elisabeth Gassiat (2009)} : Variable 	selection in model-based clustering using multilocus genotype data. Adv Data Anal Classif, Vol 3, number 2, 109-134.
	}
}

\keyword{methods}
\keyword{Genotype}
\keyword{Multi-Locus}
\keyword{Population structure}

