% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClustMMDD.R
\docType{data}
\name{genotype1}
\alias{genotype1}
\title{\code{genotype1} is a data frame of genotype data with \code{ploidy = 2.}}

\format{The format is: chr [1:1000, 1:10] "109107" "105101" "106106" ... and the 11 th column contains integers representing
	the prior classification in \code{5} sub-populations.}
	
\description{
A simulated data frame of genotype data with \code{N = 1000} individuals genotyped at \code{P = 10} loci.
	Each string represents two alleles : \code{ploidy = 2}. For example, "109107" represents \eqn{\left\{"109", "107"\right\}}. The last column of the data frame contains integers that represent the population membership.
}
\author{
Wilson Toussile
}

\seealso{
  \code{\link{genotype2}}
}

\examples{
data(genotype1)
head(genotype1)
}

\references{
\itemize{
		\item  \href{http://projecteuclid.org/euclid.ejs/1379596773}{Dominique Bontemps and Wilson Toussile (2013)} : Clustering and 	variable selection for categorical multivariate data. Electronic Journal of Statistics, Volume 7, 2344-2371, ISSN.
		
		\item 	\href{http://link.springer.com/article/10.1007\%2Fs11634-009-0043-x}{Wilson Toussile and Elisabeth Gassiat (2009)} : Variable 	selection in model-based clustering using multilocus genotype data. Adv Data Anal Classif, Vol 3, number 2, 109-134.
	}
}
\keyword{Genotype}
\keyword{Multi-Locus}
\keyword{Population structure}

