\name{read.modelKS-methods}
\docType{methods}

\alias{read.modelKS-methods}
\alias{read.modelKS}

\title{
  Read the parameters of a model \eqn{\left(K,S\right)} from a file.
}

\section{Methods}{
\describe{
\item{\code{signature()}}{Generic.}
\item{\code{signature(file = "character")}}{Read a set of parameters of a model \eqn{(K,S)} from a file.}
}}

\description{
Read the parameters of a model \eqn{\left(K,S\right)} from a file, and return an instance of \code{\linkS4class{modelKS}}.
}


\seealso{
  \code{\linkS4class{modelKS}}, \code{\link{slotNames}}, \code{\link{new}}, \code{\link{methods}}, \code{\link{show}}
}

\author{
  Wilson Toussile.
}

\keyword{methods}

