\name{setModelKS-methods}
\docType{methods}

\alias{setModelKS-methods,list-methods}
\alias{setModelKS,list-methods}

\title{
Set an instance of class \code{\linkS4class{modelKS}} from a list.
}

\arguments{
  \item{x}{A list from which to retrieve the slots of \code{\linkS4class{modelKS}}.}
}
\value{
  An object of class \code{\linkS4class{modelKS}}.
}

\description{
  Set an object of class \code{\linkS4class{modelKS}} from a list.
}

\section{Warning}{
This function is not available for users.
}

\author{
  Wilson Toussile
}

