\name{show-methods}
\alias{show}
\alias{show-methods}

\title{
  \code{show} method for an object of class \code{\linkS4class{modelKS}}
}


\arguments{
  \item{this}{An object of class \code{\linkS4class{modelKS}}.}
}

\description{
  Show an object of class \code{\linkS4class{modelKS}}.
}

\seealso{
  \code{\link{slotNames}}, \code{\link{new}}, \code{\link{methods}}, \code{\link{show}}
}

\author{
  Wilson Toussile.
}

